+++ designsafe/apps/api/external_resources/dropbox/models/files.py
@@ -82,7 +82,7 @@ class DropboxFile(object):
 
     @property
     def ext(self):
-        return os.path.splitext(self.name)[1]
+        return os.path.splitext(self.name)[1].lower()
 
     @property
     def trail(self):