+++ designsafe/apps/licenses/admin.py
@@ -4,4 +4,8 @@ import models
 
 @admin.register(models.MATLABLicense)
 class MATLABLicenseAdmin(admin.ModelAdmin):
-    pass
+    readonly_fields = ('license_type', )
+
+@admin.register(models.LSDYNALicense)
+class LSDYNALicenseAdmin(admin.ModelAdmin):
+    readonly_fields = ('license_type', )