+++ designsafe/apps/licenses/models.py
@@ -4,9 +4,10 @@ import logging
 
 logger = logging.getLogger(__name__)
 
-LICENSE_TYPES = (
-    ('MATLAB', 'MATLAB'),
-)
+LICENSE_TYPES = [
+    'MATLAB',
+    'LS-DYNA'
+]
 
 
 def get_license_info():
@@ -15,27 +16,20 @@ def get_license_info():
             'license_type': 'MATLAB',
             'class': 'designsafe.apps.licenses.MATLABLicense',
             'details_html': 'designsafe/apps/licenses/matlab_details.html',
+        },
+        {
+            'license_type': 'LS-DYNA',
+            'class': 'designsafe.apps.licenses.LSDYNALicense',
+            'details_html': 'designsafe/apps/licenses/ls-dyna_details.html'
         }
-    ]
+    ], BaseLicense.__subclasses__()
 
 
 class BaseLicense(models.Model):
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, related_name='licenses')
-    license_type = models.CharField(max_length=255, choices=LICENSE_TYPES)
+    user = models.OneToOneField(settings.AUTH_USER_MODEL, related_name='%(class)s')
 
     class Meta:
         abstract = True
-
-    def license_as_str(self):
-        return ''
-
-
-class MATLABLicense(BaseLicense):
-    license_file_content = models.TextField(help_text='This should be entire contents of'
-                                                      'the user\'s MATLAB license file. '
-                                                      'Please ensure you paste the '
-                                                      'license exactly as it is in the '
-                                                      'license file.')
 
     def __str__(self):
         return unicode(self).encode('utf-8')
@@ -46,3 +40,23 @@ class MATLABLicense(BaseLicense):
     def license_as_str(self):
         self.license_file_content = self.license_file_content.replace('\r\n', '\n')
         return self.license_file_content.encode('utf-8')
+
+
+class MATLABLicense(BaseLicense):
+    license_file_content = models.TextField(help_text='This should be entire contents of '
+                                                      'the user\'s MATLAB license file. '
+                                                      'Please ensure you paste the '
+                                                      'license exactly as it is in the '
+                                                      'license file.')
+
+    license_type = 'MATLAB'
+
+
+class LSDYNALicense(BaseLicense):
+    license_file_content = models.TextField(help_text='This should be entire contents of '
+                                                      'the user\'s LS-DYNA license file. '
+                                                      'Please ensure you paste the '
+                                                      'license exactly as it is in the '
+                                                      'license file.')
+                                                      
+    license_type = 'LS-DYNA'