+++ designsafe/apps/workspace/views.py
@@ -7,7 +7,8 @@ from django.core.urlresolvers import reverse
 from django.http import HttpResponse
 from designsafe.apps.api.notifications.models import Notification
 from designsafe.apps.workspace.tasks import JobSubmitError, submit_job
-from designsafe.apps.licenses.models import LICENSE_TYPES
+from designsafe.apps.licenses.models import LICENSE_TYPES, get_license_info
+from designsafe.libs.common.decorators import profile as profile_fn
 from requests import HTTPError
 from urlparse import urlparse
 from datetime import datetime
@@ -26,11 +27,12 @@ def index(request):
 
 
 def _app_license_type(app_id):
-    app_lic_type = app_id.split('-')[0].upper()
-    lic_type = next((t[0] for t in LICENSE_TYPES if t[0] == app_lic_type), None)
+    app_lic_type = app_id.replace('-{}'.format(app_id.split('-')[-1]), '').upper()
+    lic_type = next((t for t in LICENSE_TYPES if t in app_lic_type), None)
     return lic_type
 
 
+@profile_fn
 @login_required
 def call_api(request, service):
     try:
@@ -44,7 +46,9 @@ def call_api(request, service):
                     'type': lic_type
                 }
                 if lic_type is not None:
-                    lic = request.user.licenses.filter(license_type=lic_type).first()
+                    _, license_models = get_license_info()
+                    license_model = filter(lambda x: x.license_type == lic_type, license_models)[0]
+                    lic = license_model.objects.filter(user=request.user).first()
                     data['license']['enabled'] = lic is not None
 
             else:
@@ -70,7 +74,9 @@ def call_api(request, service):
                         'type': lic_type
                     }
                     if lic_type is not None:
-                        lic = request.user.licenses.filter(license_type=lic_type).first()
+                        _, license_models = get_license_info()
+                        license_model = filter(lambda x: x.license_type == lic_type, license_models)[0]
+                        lic = license_model.objects.filter(user=request.user).first()
                         data['license']['enabled'] = lic is not None
 
                 else:
@@ -133,7 +139,9 @@ def call_api(request, service):
                     # check for running licensed apps
                     lic_type = _app_license_type(job_post['appId'])
                     if lic_type is not None:
-                        lic = request.user.licenses.filter(license_type=lic_type).first()
+                        _, license_models = get_license_info()
+                        license_model = filter(lambda x: x.license_type == lic_type, license_models)[0]
+                        lic = license_model.objects.filter(user=request.user).first()
                         job_post['parameters']['_license'] = lic.license_as_str()
 
                     # url encode inputs