+++ designsafe/static/scripts/data-depot/controllers/projects.js
@@ -222,19 +222,20 @@
       DataBrowserService.showListing();
     };
 
-    $scope.showPreview = function(){
-      //DataBrowserService.state().showMainListing = false;
-      //DataBrowserService.state().showPreviewListing = true;
-      $scope.previewHref = undefined;
-      DataBrowserService.showPreview();
-      FileListing.get({'system': $scope.browser.listing.system,
-                       'name': 'projectimage.jpg',
-                       'path': '/projectimage.jpg'}).then(function(list){
-                        list.preview().then(function(data){
-                            $scope.previewHref = data.postit;
-                        });
-                      });
-    };
+    // disabling project image feature
+    // $scope.showPreview = function(){
+    //   //DataBrowserService.state().showMainListing = false;
+    //   //DataBrowserService.state().showPreviewListing = true;
+    //   $scope.previewHref = undefined;
+    //   DataBrowserService.showPreview();
+    //   FileListing.get({'system': $scope.browser.listing.system,
+    //                    'name': 'projectimage.jpg',
+    //                    'path': '/projectimage.jpg'}).then(function(list){
+    //                     list.preview().then(function(data){
+    //                         $scope.previewHref = data.postit;
+    //                     });
+    //                   });
+    // };
 
     function savePublication(){
       var publication = angular.copy($scope.state.publication);